Refactor all logic out of +page.server.ts
