+++ src/lib/types/discography/extra.ts
@@ -1,0 +1,9 @@
+import type { DiscographyObject } from "./entries";
+
+export interface Extra {
+    "lastUpdate": Date
+}
+
+export interface ExtraObject extends DiscographyObject {
+    "extra": Extra
+}
